requirements.txt
fastapi
pytesseract
uvicorn
Pillow
python-multipart